--- packages/react-native-aria/interactions/src/index.web.ts
@@ -1,7 +1,0 @@
-export { useHover } from './useHover.web';
-export { usePress } from './usePress';
-export {
-  keyboardDismissHandlerManager,
-  useKeyboardDismissable,
-  useBackHandler,
-} from './useKeyboardDismisssable';
