+++ packages/react-native-aria/interactions/src/useKeyboardDismisssable.ts
@@ -1,6 +1,5 @@
-import React from 'react';
 import { useEffect } from 'react';
-import { BackHandler, Platform } from 'react-native';
+import { useBackHandler } from './useBackHandler';
 
 type IParams = {
   enabled?: boolean;
@@ -27,7 +26,7 @@ export const keyboardDismissHandlerManager = {
  * Handles attaching callback for Escape key listener on web and Back button listener on Android
  */
 export const useKeyboardDismissable = ({ enabled, callback }: IParams) => {
-  React.useEffect(() => {
+  useEffect(() => {
     let cleanupFn = () => {};
     if (enabled) {
       cleanupFn = keyboardDismissHandlerManager.push(callback);
@@ -41,32 +40,3 @@ export const useKeyboardDismissable = ({ enabled, callback }: IParams) => {
 
   useBackHandler({ enabled, callback });
 };
-
-export function useBackHandler({ enabled, callback }: IParams) {
-  useEffect(() => {
-    if (Platform.OS === 'web') {
-      const handleEscape = (e: KeyboardEvent) => {
-        if (e.key === 'Escape') {
-          callback();
-        }
-      };
-
-      document?.body?.addEventListener?.('keyup', handleEscape);
-      return () => {
-        document?.body?.removeEventListener?.('keyup', handleEscape);
-      };
-    } else {
-      let backHandler = () => {
-        callback();
-        return true;
-      };
-      if (enabled) {
-        BackHandler.addEventListener('hardwareBackPress', backHandler);
-      } else {
-        BackHandler.removeEventListener('hardwareBackPress', backHandler);
-      }
-      return () =>
-        BackHandler.removeEventListener('hardwareBackPress', backHandler);
-    }
-  }, [enabled, callback]);
-}
